+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/enums/ModeType.java
@@ -1,0 +1,5 @@
+package de.gematik.test.tiger.lib.enums;
+
+public enum ModeType {
+    XML, JSON
+}